requirements.txt
passlib[bcrypt]
python-jose[cryptography]
fastapi[all]
sqlalchemy
python-multipart
pydantic[email]
aiosqlite
pandas
requests
python-decouple==3.8